chore(hooks): consolidate websocket helpers into useAgentState; remove legacy useAgentStateWebSocket.ts
